Move dtt to torchscript_operators to highlight Torch isolation
