+++ polymetis/polymetis/src/torchscript_operators/pinocchio.cpp
@@ -6,7 +6,7 @@
 #include <iostream>
 #include <string>
 
-#include "dtt.h"
+#include "polymetis/torchscript_operators/dtt.h"
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
